Refactor: update map type symbol class and draw logic
